+++ src/pretalx/api/serializers/submission.py
@@ -3,9 +3,10 @@ from rest_framework.serializers import (
     ModelSerializer, SerializerMethodField, SlugRelatedField,
 )
 
+from pretalx.api.serializers.question import AnswerSerializer
 from pretalx.api.serializers.speaker import SubmitterSerializer
 from pretalx.schedule.models import Schedule, TalkSlot
-from pretalx.submission.models import Submission, SubmissionStates
+from pretalx.submission.models import Answer, Submission, SubmissionStates
 
 
 class SlotSerializer(I18nAwareModelSerializer):
@@ -13,9 +14,7 @@ class SlotSerializer(I18nAwareModelSerializer):
 
     class Meta:
         model = TalkSlot
-        fields = (
-            'room', 'start', 'end',
-        )
+        fields = ('room', 'start', 'end')
 
 
 class SubmissionSerializer(I18nAwareModelSerializer):
@@ -23,15 +22,40 @@ class SubmissionSerializer(I18nAwareModelSerializer):
     submission_type = SlugRelatedField(slug_field='name', read_only=True)
     slot = SlotSerializer(TalkSlot.objects.filter(is_visible=True), read_only=True)
     duration = SerializerMethodField()
+    answers = SerializerMethodField()
+
+    @property
+    def is_orga(self):
+        request = self.context.get('request')
+        if request:
+            return request.user.has_perm('orga.view_submissions', request.event)
 
     def get_duration(self, obj):
         return obj.export_duration
 
+    def get_answers(self, obj):
+        if self.is_orga:
+            return AnswerSerializer(
+                Answer.objects.filter(submission=obj), many=True
+            ).data
+        return []
+
     class Meta:
         model = Submission
         fields = (
-            'code', 'speakers', 'title', 'submission_type', 'state', 'abstract',
-            'description', 'duration', 'do_not_record', 'content_locale', 'slot', 'image',
+            'code',
+            'speakers',
+            'title',
+            'submission_type',
+            'state',
+            'abstract',
+            'description',
+            'duration',
+            'do_not_record',
+            'content_locale',
+            'slot',
+            'image',
+            'answers',
         )
 
 
@@ -43,16 +67,14 @@ class ScheduleListSerializer(ModelSerializer):
 
     class Meta:
         model = Schedule
-        fields = (
-            'version',
-        )
+        fields = ('version',)
 
 
 class ScheduleSerializer(ModelSerializer):
-    slots = SubmissionSerializer(Submission.objects.filter(state=SubmissionStates.CONFIRMED), many=True)
+    slots = SubmissionSerializer(
+        Submission.objects.filter(state=SubmissionStates.CONFIRMED), many=True
+    )
 
     class Meta:
         model = Schedule
-        fields = (
-            'slots', 'version',
-        )
+        fields = ('slots', 'version')